Nts working crud

+++ app/Http/Controllers/OrgController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use App\Models\Org;
 use App\Http\Controllers\Controller;
 use Symfony\Component\HttpFoundation\Response;
@@ -8,6 +9,9 @@ use Illuminate\Http\Request;
 use App\Http\Requests\StoreOrgRequest;
 use App\Http\Requests\MassDestroyOrgRequest;
 use App\Http\Requests\UpdateOrgRequest;
+use App\Models\PaymentPlan;
+use Illuminate\Contracts\Database\Eloquent\Builder;
+
 class OrgController extends Controller
 {
     /**
@@ -20,9 +24,17 @@ class OrgController extends Controller
         abort_if(!auth()->user()->can('read-org'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $inputSearchString = $request->input('s', '');
-        $orgs = Org::with('media')->when($inputSearchString, function($query) use ($inputSearchString) {
-                $query->where(function($query) use ($inputSearchString) {
-                    $query->orWhere('name', 'LIKE', '%'.$inputSearchString.'%');
+
+        $orgs = Org::query()
+            ->with('plan:id,plan_name')
+            ->when($inputSearchString, function ($query) use ($inputSearchString) {
+                $query->where(function ($query) use ($inputSearchString) {
+                    $query->orWhere('name', 'LIKE', '%' . $inputSearchString . '%');
+                    $query->orWhere(function ($query) use ($inputSearchString) {
+                        $query->whereHas('plan', function (Builder $builder) use ($inputSearchString) {
+                            $builder->where('plan_name', 'LIKE', '%' . $inputSearchString . '%');
+                        });
+                    });
                 });
             })
             ->isActive()
@@ -30,9 +42,7 @@ class OrgController extends Controller
             ->paginate(config('app-config.datatable_default_row_count', 25))
             ->withQueryString();
 
-        return view('orgs.index', [
-            'orgs' => $orgs
-        ]);
+        return view('orgs.index', compact('orgs'));
     }
 
     /**
@@ -44,7 +54,13 @@ class OrgController extends Controller
     {
         abort_if(!auth()->user()->can('create-org'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        return view('orgs.create');
+        $plans = PaymentPlan::query()
+            ->select('id', 'plan_name')
+            ->isActive()
+            ->orderBy('plan_name')
+            ->get();
+
+        return view('orgs.create', compact('plans'));
     }
 
     /**
@@ -55,13 +71,15 @@ class OrgController extends Controller
      */
     public function store(StoreOrgRequest $request)
     {
+        abort_if(!auth()->user()->can('create-org'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $org = Org::create($request->safe()->only('name', 'plan_id', 'address', 'is_premium'));
 
-        if($request->hasFile('attachments')){
+        if ($request->hasFile('attachments')) {
             $org->addMultipleMediaFromRequest(['attachments'])
-            ->each(function($attachment){
-                $attachment->toMediaCollection('attachments');
-            });
+                ->each(function ($attachment) {
+                    $attachment->toMediaCollection('attachments');
+                });
         }
         toast(__('global.crud_actions', ['module' => 'Org', 'action' => 'created']), 'success');
 
@@ -78,6 +96,9 @@ class OrgController extends Controller
     {
         abort_if(!auth()->user()->can('show-org'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
+        $org->load('plan:id,plan_name');
+        $org->load('media');
+
         return view('orgs.show', compact('org'));
     }
 
@@ -89,7 +110,15 @@ class OrgController extends Controller
      */
     public function edit(Org $org)
     {
-        return view('orgs.edit', compact('org'));
+        abort_if(!auth()->user()->can('update-org'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        $plans = PaymentPlan::query()
+            ->select('id', 'plan_name')
+            ->isActive()
+            ->orderBy('plan_name')
+            ->get();
+
+        return view('orgs.edit', compact('org', 'plans'));
     }
 
     /**
@@ -101,14 +130,16 @@ class OrgController extends Controller
      */
     public function update(UpdateOrgRequest $request, Org $org)
     {
-        $org->update($request->safe()->only(['name','plan_id', 'address', 'is_premium']));
+        abort_if(!auth()->user()->can('update-org'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        if($request->hasFile('attachments')){
+        $org->update($request->safe()->only(['name', 'plan_id', 'address', 'is_premium']));
+
+        if ($request->hasFile('attachments')) {
             $org->media()->delete();
             $org->addMultipleMediaFromRequest(['attachments'])
-            ->each(function($attachment){
-                $attachment->toMediaCollection('attachments');
-            });
+                ->each(function ($attachment) {
+                    $attachment->toMediaCollection('attachments');
+                });
         }
         toast(__('global.crud_actions', ['module' => 'Org', 'action' => 'updated']), 'success');
 
@@ -139,6 +170,8 @@ class OrgController extends Controller
                 'is_active' => 3,
                 'updatedby_userid' => auth()->id(),
             ]);
+
+        toast(__('global.crud_actions', ['module' => 'Org', 'action' => 'deleted']), 'success');
 
         return response(null, Response::HTTP_NO_CONTENT);
     }
